feat: 防災タスク関係達成カラムをbyteの上限が足りないため、int型からbigint型に変更 #78

+++ api/db/migrate/20220723003748_create_task_profiles.rb
@@ -2,9 +2,9 @@ class CreateTaskProfiles < ActiveRecord::Migration[6.1]
   def change
     create_table :task_profiles do |t|
       t.references :user,     null: false, foreign_key: true, comment: 'ユーザーID'
-      t.integer :sns_tasks,   null: false, default: 0,        comment: 'SNS関係防災タスク'
-      t.integer :house_tasks, null: false, default: 0,        comment: '家庭関係防災タスク'
-      t.integer :stock_tasks, null: false, default: 0,        comment: '備蓄関係防災タスク'
+      t.bigint :sns_tasks,    null: false, default: 0,        comment: 'SNS関係防災タスク'
+      t.bigint :house_tasks,  null: false, default: 0,        comment: '家庭関係防災タスク'
+      t.bigint :stock_tasks,  null: false, default: 0,        comment: '備蓄関係防災タスク'
       t.timestamps
     end
   end
